requirements.txt
requests~=2.32.5
python-dotenv~=1.2.2
azure-identity~=1.17.0
azure-keyvault-secrets~=4.8.0
azure-storage-blob
azure-functions
